배터리 충전 그래프 UI 추가
