+++ EZMacker/Service/Enum/AppSmartProcessInfo.swift
@@ -1,0 +1,13 @@
+//
+//  AppSmartProcessInfo.swift
+//  EZMacker
+//
+//  Created by 박유경 on 5/26/24.
+//
+
+import Foundation
+struct AppSmartProcessInfo {
+    let pid: Int32
+    let name: String
+    let cpuUsage: Double
+}
