+++ EZMacker/ViewModel/SmartBatteryViewModel.swift
@@ -7,18 +7,6 @@
 import Combine
 import SwiftUI
 
-struct ChargeData: Identifiable {
-    var id = UUID()
-    var vacVoltageLimit: Int
-    var chargingCurrent: CGFloat
-    var timeChargingThermallyLimited: Int
-    var chargerStatus: Data
-    var chargingVoltage: CGFloat
-    var chargerInhibitReason: Int
-    var chargerID: Int
-    var notChargingReason: Int
-}
-
 class SmartBatteryViewModel<ProvidableType: AppSmartBatteryRegistryProvidable>: ObservableObject {
     
     deinit {
@@ -133,7 +121,7 @@ extension SmartBatteryViewModel {
                       return nil
                   }
                   return ChargeData(vacVoltageLimit: vacVoltageLimit,
-                                    chargingCurrent: 700,
+                                    chargingCurrent: chargingCurrent,
                                     timeChargingThermallyLimited: timeChargingThermallyLimited,
                                     chargerStatus: chargerStatus,
                                     chargingVoltage: chargingVoltage,
@@ -142,13 +130,12 @@ extension SmartBatteryViewModel {
                                     notChargingReason: notChargingReason)
               }
               .receive(on: DispatchQueue.main)
-              .sink { [weak self] chargeData in
+              .sink { [weak self] charge in
                   guard let self = self else { return }
-                  self.chargeData.append(chargeData)
-                  if self.chargeData.count > 6 {
-                      self.chargeData.removeFirst()
+                  if chargeData.count > 5 {
+                      self.chargeData.removeAll()
                   }
-                  
+                  chargeData.append(charge)
               }
               .store(in: &cancellables)
         
@@ -233,159 +220,49 @@ extension SmartBatteryViewModel {
         systemPreferenceService.openSystemPreferences(systemPath: settingPath)
     }
 }
-struct ProcessInfo {
-    let pid: Int32
-    let name: String
-    let cpuUsage: Double
-}
 
 
 extension SmartBatteryViewModel {
-    func runShellCommand(_ command: String) -> String? {
-        let process = Process()
-        process.launchPath = "/bin/bash"
-        process.arguments = ["-c", command]
-
-        let pipe = Pipe()
-        process.standardOutput = pipe
-        process.standardError = pipe
-
-        let fileHandle = pipe.fileHandleForReading
-        process.launch()
-
-        let data = fileHandle.readDataToEndOfFile()
-        process.waitUntilExit()
-
-        let output = String(data: data, encoding: .utf8)
-        return output
-    }
-
-    // Function to get the list of processes with high CPU usage
-    func getHighCpuProcesses() -> [(pid: String, name: String, cpu: String)] {
-        guard let output = runShellCommand("ps aux | sort -nrk 3 | head -10") else {
-            return []
-        }
-
-        var processes: [(pid: String, name: String, cpu: String)] = []
-        let lines = output.split(separator: "\n").dropFirst()
-
-        for line in lines {
-            let columns = line.split(separator: " ", omittingEmptySubsequences: true)
-            if columns.count >= 11 {
-                let pid = String(columns[1])
-                let cpu = String(columns[2])
-                let name = columns[10...].joined(separator: " ")
-                processes.append((pid: pid, name: name, cpu: cpu))
-            }
-        }
-
-        return processes
-    }
-
-    func hostCPULoadInfo() -> host_cpu_load_info? {
-        let HOST_CPU_LOAD_INFO_COUNT = MemoryLayout<host_cpu_load_info>.stride/MemoryLayout<integer_t>.stride
-        var size = mach_msg_type_number_t(HOST_CPU_LOAD_INFO_COUNT)
-        var cpuLoadInfo = host_cpu_load_info()
-
-        let result = withUnsafeMutablePointer(to: &cpuLoadInfo) {
-            $0.withMemoryRebound(to: integer_t.self, capacity: HOST_CPU_LOAD_INFO_COUNT) {
-                host_statistics(mach_host_self(), HOST_CPU_LOAD_INFO, $0, &size)
-            }
-        }
-        if result != KERN_SUCCESS{
-            print("Error  - \(#file): \(#function) - kern_result_t = \(result)")
-            return nil
-        }
-        return cpuLoadInfo
-    }
-
-
-}
-class MyCpuUsage {
-    var cpuInfo: processor_info_array_t!
-    var prevCpuInfo: processor_info_array_t?
-    var numCpuInfo: mach_msg_type_number_t = 0
-    var numPrevCpuInfo: mach_msg_type_number_t = 0
-    let numCPUs: uint
-    let CPUUsageLock = NSLock()
-
-    init() {
-        // obtaining numCPUs
-        let mibKeys: [Int32] = [ CTL_HW, HW_NCPU ]
-        var numCPUs: uint = 0
-        mibKeys.withUnsafeBufferPointer() { mib in
-            var sizeOfNumCPUs = MemoryLayout<uint>.size
-            let status = sysctl(processor_info_array_t(mutating: mib.baseAddress), 2, &numCPUs, &sizeOfNumCPUs, nil, 0)
-            if status != 0 {
-                numCPUs = 1
-            }
-        }
-        self.numCPUs = numCPUs
-
-        // calling the updateInfo() func to initialize prevCpuInfo
-        updateInfo()
-        // sleep 1 second
-        sleep(1)
-        // calling the updateInfo() func to print cpu usage during past second
-        updateInfo()
-    }
-
-    func updateInfo() {
-        var numCPUsU: natural_t = 0
-        let err: kern_return_t = host_processor_info(mach_host_self(), PROCESSOR_CPU_LOAD_INFO, &numCPUsU, &cpuInfo, &numCpuInfo);
-        guard err == KERN_SUCCESS else {
-            print("Error host_processor_info!")
-            return
-        }
-
-        // Two variables to calculate sum of all cores
-        var totalInUse: Int32 = 0
-        var totalTotal: Int32 = 0
-
-        CPUUsageLock.lock()
-
-        for i in 0 ..< Int32(numCPUs) {
-            var inUse: Int32
-            var total: Int32
-            if let prevCpuInfo = prevCpuInfo {
-                inUse = cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_USER)]
-                    - prevCpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_USER)]
-                    + cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_SYSTEM)]
-                    - prevCpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_SYSTEM)]
-                    + cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_NICE)]
-                    - prevCpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_NICE)]
-                total = inUse + (cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_IDLE)]
-                    - prevCpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_IDLE)])
-            } else {
-                inUse = cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_USER)]
-                    + cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_SYSTEM)]
-                    + cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_NICE)]
-                total = inUse + cpuInfo[Int(CPU_STATE_MAX * i + CPU_STATE_IDLE)]
-            }
-
-            // Assigning totals here
-            totalInUse += inUse
-            totalTotal += total
-        }
-        CPUUsageLock.unlock()
-
-        if let prevCpuInfo = prevCpuInfo {
-            // I free the memory of prevCpuInfo
-            let prevCpuInfoSize: size_t = MemoryLayout<integer_t>.stride * Int(numPrevCpuInfo)
-            vm_deallocate(mach_task_self_, vm_address_t(bitPattern: prevCpuInfo), vm_size_t(prevCpuInfoSize))
-            
-            // I print sum of all cores percentage.
-            let totalUsagePercentage = Float(totalInUse) / Float(totalTotal) * 100
-            print(String(format: "Total CPU Usage: %.2f%%", totalUsagePercentage))
-        }
-
-        prevCpuInfo = cpuInfo
-        numPrevCpuInfo = numCpuInfo
-
-        cpuInfo = nil
-        numCpuInfo = 0
-    }
-    
+//    func runShellCommand(_ command: String) -> String? {
+//        let process = Process()
+//        process.launchPath = "/bin/bash"
+//        process.arguments = ["-c", command]
+//
+//        let pipe = Pipe()
+//        process.standardOutput = pipe
+//        process.standardError = pipe
+//
+//        let fileHandle = pipe.fileHandleForReading
+//        process.launch()
+//
+//        let data = fileHandle.readDataToEndOfFile()
+//        process.waitUntilExit()
+//
+//        let output = String(data: data, encoding: .utf8)
+//        return output
+//    }
+//
+//    func getHighCpuProcesses() -> [(pid: String, name: String, cpu: String)] {
+//        guard let output = runShellCommand("ps aux | sort -nrk 3 | head -10") else {
+//            return []
+//        }
+//
+//        var processes: [(pid: String, name: String, cpu: String)] = []
+//        let lines = output.split(separator: "\n").dropFirst()
+//
+//        for line in lines {
+//            let columns = line.split(separator: " ", omittingEmptySubsequences: true)
+//            if columns.count >= 11 {
+//                let pid = String(columns[1])
+//                let cpu = String(columns[2])
+//                let name = columns[10...].joined(separator: " ")
+//                processes.append((pid: pid, name: name, cpu: cpu))
+//            }
+//        }
+//
+//        return processes
+//    }
+//
 //    func hostCPULoadInfo() -> host_cpu_load_info? {
 //        let HOST_CPU_LOAD_INFO_COUNT = MemoryLayout<host_cpu_load_info>.stride/MemoryLayout<integer_t>.stride
 //        var size = mach_msg_type_number_t(HOST_CPU_LOAD_INFO_COUNT)
@@ -403,3 +280,4 @@ class MyCpuUsage {
 //        return cpuLoadInfo
 //    }
 }
+
